statistics-api/requirements.txt

# Python dependencies for the Statistics API

fastapi==0.111.0        # Web framework — handles routing, validation, serialization
uvicorn==0.30.1         # ASGI server — runs the FastAPI app
httpx==0.27.0           # Async HTTP client — used to call the Device Registration API
psycopg2-binary==2.9.9  # PostgreSQL driver — used for the statistics query
pydantic==2.7.1         # Data validation — used by FastAPI for request/response models